+++ apps/docs-snippets2/sway/Forc.toml
@@ -16,5 +16,6 @@ members = [
   "script-sum",
   "simple-predicate",
   "sum-option-u8",
-  "return-context"
+  "return-context",
+  "transfer-to-address"
 ]
